backend/requirements.txt
# ============================================================================
# DOG WRITER MVP - STREAMLINED REQUIREMENTS (Production Ready)
# ============================================================================

# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.18
python-dotenv==1.0.0

# Authentication & Security
PyJWT==2.8.0
bcrypt==4.2.1
email-validator==2.2.0
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
cryptography>=42.0.2

# HTTP requests & async support
requests==2.32.4
aiohttp==3.10.11
aiofiles==23.2.1

# AI integrations (Core MVP features only)
openai>=1.35.0
google-generativeai>=0.8.5

# Database (PostgreSQL only - production ready)
psycopg2-binary==2.9.9

# Templating (for email/HTML if needed)
jinja2==3.1.6

# Performance monitoring (lightweight)
psutil==6.1.0
